PairComp/requirements.txt
apex==0.9.10dev
datasets==3.6.0
decord==0.6.0
deepspeed==0.16.4
einops==0.8.1
fastchat==0.1.0
flash_attn==2.7.4.post1
imageio==2.37.0
numpy==2.3.0
opencv_python==4.11.0.86
orjson==3.10.18
packaging==25.0
peft==0.10.0
Pillow==11.2.1
sentencepiece==0.1.99
timm==0.9.12
tokenizers==0.21.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==4.37.2
